chore: move more classes into `proxycore.listeners`
