--- vane-proxy-core/src/main/java/org/oddlama/vane/proxycore/ProxyEvent.java
@@ -1,9 +1,0 @@
-package org.oddlama.vane.proxycore;
-
-public interface ProxyEvent {
-
-	ProxyPendingConnection get_connection();
-
-	void fire();
-
-}
