+++ vane-proxy-core/src/main/java/org/oddlama/vane/proxycore/listeners/ProxyEvent.java
@@ -1,0 +1,11 @@
+package org.oddlama.vane.proxycore.listeners;
+
+import org.oddlama.vane.proxycore.ProxyPendingConnection;
+
+public interface ProxyEvent {
+
+	ProxyPendingConnection get_connection();
+
+	void fire();
+
+}
