+++ vane-waterfall/src/main/java/org/oddlama/vane/waterfall/compat/BungeeCompatProxyServerPing.java
@@ -2,7 +2,7 @@ package org.oddlama.vane.waterfall.compat;
 
 import net.md_5.bungee.api.ServerPing;
 import net.md_5.bungee.api.chat.TextComponent;
-import org.oddlama.vane.proxycore.ProxyServerPing;
+import org.oddlama.vane.proxycore.listeners.ProxyServerPing;
 
 public class BungeeCompatProxyServerPing implements ProxyServerPing {
